tagging/tagging_requirements.txt
lancedb
librosa
google-generativeai
python-dotenv
pandas
numpy
soundfile